+++ api/CHANGELOG.md
@@ -9,6 +9,8 @@
 - Improved patcher logging/profiling and error surfacing
 - Improved metadata reading for split APKs on the app info page
 - Imrpoved metedata reading for regular APKs on the app info page
+- Converted the `Save patched app` button, `Export` button on the `App info` screen for saved patched apps, and the `Export` button on the Download settings page to use the custom file picker
+- Added a saving modal to the custom file picker
 
 
 # Bug fixes